Moving tests back to test_profile_plots
